+++ src/components/elements/Picture/types.d.ts
@@ -1,0 +1,6 @@
+import { GlobProps } from "../../../typings";
+
+export interface PictureProps extends GlobProps, HTMLPictureElement {
+  dataSrc?: string;
+  srcSets?: HTMLSourceElement[];
+}